refactor: update material components imports to use shadcn directory
