+++ features/materials/components/material-unit-selector.tsx
@@ -6,7 +6,7 @@ import {
   SelectLabel,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/components/shadcn/select';
 
 export const MATERIAL_UNITS: {
   group: string;
